fix: fellowship fixes
